requirements.txt
openai
pyyaml
pillow
pydantic
matplotlib
pyqt6
pandas
pyarrow